refactor: gbif module for CLI
