+++ crumbs/gbif/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for accessing GBIF occurrences records
+"""
