+++ crumbs/gbif/cli/__init__.py
@@ -1,0 +1,3 @@
+"""
+Command Line Interface for accessing GBIF occurrences records
+"""
