+++ setup.cfg
@@ -27,5 +27,5 @@ exclude =
 
 [options.entry_points]
 console_scripts =
-    crumbs-get-gbif = crumbs.get_gbif:main
+    crumbs-get-gbif = crumbs.gbif.cli.request_cmd:main
     crumbs-fit-sdm = crumbs.sdm.fit:main
